volume is from 0.0-1.0

+++ util/sound.h
@@ -29,7 +29,7 @@ public:
 	Sound & operator=( const Sound & rhs );
 
 	void play();
-	void play( int volume, int pan );
+	void play(double volume, int pan);
 	void playLoop();
         void stop();
